Support monthly transfers greater than 2GB/month in the log.
svn path=/branches/KDE/3.5/kdenetwork/kppp/; revision=502587

+++ logview/loginfo.cpp
@@ -65,15 +65,15 @@ double LogInfo::totalCosts() const {
   return _total_cost;
 }
 
-int LogInfo::bytesIn() const {
+double LogInfo::bytesIn() const {
   return _bytes_in;
 }
 
-int LogInfo::bytesOut() const {
+double LogInfo::bytesOut() const {
   return _bytes_out;
 }
 
-int LogInfo::bytes() const {
+double LogInfo::bytes() const {
   if(bytesIn() == -1 || bytesOut() == -1)
     return -1;
   else
